backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.8
qdrant-client==1.6.9
tavily-python==0.3.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
aiofiles==23.2.1
google-generativeai